+++ pyobjc-framework-VideoToolbox/PyObjCTest/test_vtprofessionalvideoworkflow.py
@@ -1,0 +1,12 @@
+from PyObjCTools.TestSupport import *
+import VideoToolbox
+
+class TestVTProfessionalVideoWorkflow (TestCase):
+    @min_os_level('10.9')
+    def test_functions(self):
+        VideoToolbox.VTRegisterProfessionalVideoWorkflowVideoDecoders
+        VideoToolbox.VTRegisterProfessionalVideoWorkflowVideoEncoders
+
+
+if __name__ == "__main__":
+    main()